Text Encryption using Fernet
This is a project that uses Fernet, a symmetric encryption algorithm. It uses AES (Advanced Encryption Standard) and CBC (Cipher Block Chaining) to encrypt data. It also uses SHA256 to ensure data has not been tampered with. This project allows the user to input plaintext data, encrypt the data, decrypt the data, and save each step to a file, while also generating and being able to reuse the keys for encryption.